pyproject.toml
[project]
name = "youtube-transcripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "pytube>=15.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "youtube-transcript-api>=1.2.4",
    "yt-dlp>=2025.12.8",
]
